requirements.txt
streamlit
requests
biopython
plotly
